backend/requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
requests==2.25.1
openai
python-dotenv==0.17.1
Jinja2==2.11.3
MarkupSafe==1.1.1